requirements.txt
astroid==3.3.11
dill==0.4.0
isort==6.1.0
mccabe==0.7.0
numpy==2.2.6
opencv-python==4.12.0.88
pillow==11.3.0
platformdirs==4.4.0
pylint==3.3.9
tomlkit==0.13.3
astroid==3.3.11
contourpy==1.3.3
cycler==0.12.1
dill==0.4.0
fonttools==4.60.1
isort==6.1.0
kiwisolver==1.4.9
matplotlib==3.10.6
mccabe==0.7.0
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
platformdirs==4.4.0
pylint==3.3.9
pyparsing==3.2.5
python-dateutil==2.9.0.post0
six==1.17.0
tomlkit==0.13.3
